Refactor postprocessing module exports
